Rename Me component to Addresses in vue-starter
